requirements.txt
python-telegram-bot==13.14
redis==3.2.1
environs==9.5.0